chore(team): split step materialization helpers
